separate data-ipv6-prefix-list
