+++ catalystwan/endpoints/configuration/policy/list/data_ipv6_prefix.py
@@ -5,17 +5,13 @@ from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
-from catalystwan.models.policy.lists import DataIPv6PrefixList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.data_ipv6_prefix import (
+    DataIPv6PrefixList,
+    DataIPv6PrefixListEditPayload,
+    DataIPv6PrefixListInfo,
+)
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class DataIPv6PrefixListEditPayload(DataIPv6PrefixList, PolicyListId):
-    pass
-
-
-class DataIPv6PrefixListInfo(DataIPv6PrefixList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicyDataIPv6PrefixList(APIEndpoints, PolicyListEndpoints):
