+++ catalystwan/models/policy/__init__.py
@@ -12,6 +12,7 @@ from catalystwan.models.policy.list.as_path import ASPathList
 from catalystwan.models.policy.list.class_map import ClassMapList
 from catalystwan.models.policy.list.color import ColorList
 from catalystwan.models.policy.list.communities import CommunityList, ExpandedCommunityList
+from catalystwan.models.policy.list.data_ipv6_prefix import DataIPv6PrefixList
 
 from .centralized import CentralizedPolicy, TrafficDataDirection
 from .definition.access_control_list import AclPolicy
@@ -29,7 +30,6 @@ from .definition.traffic_data import TrafficDataPolicy
 from .definition.vpn_membership import VPNMembershipPolicy
 from .definition.zone_based_firewall import ZoneBasedFWPolicy
 from .lists import (
-    DataIPv6PrefixList,
     DataPrefixList,
     FQDNList,
     GeoLocationList,
